Working on migrating to a less hacky login system
