+++ index.php
@@ -6,7 +6,7 @@
 		<meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1, maximum-scale=1, user-scalable=0" />
 		<meta name="apple-mobile-web-app-capable" content="no" />
 		<meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />
-		
+
 		<!--favicon generated by http://realfavicongenerator.net/-->
 		<link rel="apple-touch-icon" sizes="180x180" href="/favicons/apple-touch-icon.png">
 		<link rel="icon" type="image/png" href="/favicons/favicon-32x32.png" sizes="32x32">
@@ -23,7 +23,7 @@
 <body>
 
 	<h1 id="title">Scouting <img id="hamburger" src="images/hamburger.png" height="50" /></h1>
-	
+
 	<div hidden id="hamburgermenu">
 		<ul id="menuitems">
 			<li id="searchli">Search</li>
@@ -31,12 +31,12 @@
 			<a href="edit.php"><li id="addli">Add Robot</li></a>
 		</ul>
 	</div>
-	
+
 	<div hidden id="TeamSearch">
 		<input name="TeamSearchbox" placeholder="Search for a team here!" pattern="[0-9]*" type="number">
 		<div id="closesearchbar"> X </div>
 	</div>
-	
+
 	<div hidden id="Filters">
 		<div id="driveSystemDiv">
 			<input type="checkbox" id="DriveSystemCheck">
@@ -54,7 +54,7 @@
 		}
 		?>
 	</div>
-	
+
 	<div id="container">
 	</div>
 	<?php
@@ -62,7 +62,7 @@
 		echo('<div id="loginbox"><p style="margin:2px;"><a id="logoutLink" href="logout.php">Log Out: '.$_SESSION["user"].'</a></p></div>');
 	}
 	?>
-	
+
 	<script src="scripts/jquery-3.1.1.min.js"></script>
 	<script src="scripts/updateList.js"></script>
 	<script src="scripts/hamburger.js"></script>
