+++ login.php
@@ -4,14 +4,12 @@
 		<title>Log in</title>
 		<link rel="stylesheet" href="style.css">
 		<?php
+		require 'functions.php';
+
+		$passFailed = False;
+
 		if ($_SERVER["REQUEST_METHOD"] == "GET") {
 			$redirect = $_GET['redirect'];
-		}
-		function clean($data) {
-		  $data = trim($data);
-		  $data = stripslashes($data);
-		  $data = htmlspecialchars($data);
-		  return $data;
 		}
 		if ($_SERVER["REQUEST_METHOD"] == "POST") {
 			if(isset($_POST["redirect"])) {
@@ -19,11 +17,12 @@
 			}
 			$user = clean($_POST["user"]);
 			$pass = clean($_POST["pass"]);
-			$DB = new mysqli("localhost",'momentu2_' . $user,$pass,"momentu2_frcteam4999");
 			$_SESSION["user"] = $user;
 			$_SESSION["pass"] = $pass;
-			if (!$DB->connect_error) {
+			if (checkUserPassword($user, $pass)) {
 				$_SESSION["loggedIn"] = True;
+			} else {
+				$passFailed = True;
 			}
 		}
 		if (isset($_SESSION["loggedIn"])) {
@@ -35,11 +34,11 @@
 			exit();
 		}
 		?>
-	
+
 		<meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1, maximum-scale=1, user-scalable=0" />
 		<meta name="apple-mobile-web-app-capable" content="no" />
 		<meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />
-		
+
 		<!--favicon generated by http://realfavicongenerator.net/-->
 		<link rel="apple-touch-icon" sizes="180x180" href="/favicons/apple-touch-icon.png">
 		<link rel="icon" type="image/png" href="/favicons/favicon-32x32.png" sizes="32x32">
@@ -54,19 +53,14 @@
 		<meta name="theme-color" content="#ffffff">
 	</head>
 	<body>
-		<form id="loginform" action="<?php echo htmlspecialchars($_SERVER["PHP_SELF"]);?>" method="post">
+		<form id="loginform" action="<?php echo(htmlspecialchars($_SERVER["PHP_SELF"]));?>" method="post">
 			<p id="userlabel">Username:</p><br>
 			<input id="usernamefield" type="text" name="user" <?php if($_SERVER["REQUEST_METHOD"] == "POST"){echo('value="'.$user.'"');}?>><br>
 			<p id="passlabel">Password:</p><br>
 			<input id="passwordfield" type="password" name="pass"><br>
 			<?php
-			if (isset($DB) and $DB->connect_error) {
-				$error = $DB->connect_error;
-				if(is_int(strpos($error, "Access denied"))) {
-					echo('<p style="font-size: 25px; color: red;">Access Denied: Check Username and Password</p>');
-				} else {
-					echo('<p sytle = "font-size: 15px; font-color: red;">'.$error.'</p>');
-				}
+			if ($passFailed) {
+				echo('<p style="font-size: 25px; color: red;">Access Denied: Check Username and Password</p>');
 			}
 			if(isset($redirect)){
 				echo('<input type="hidden" name="redirect" value="'.$redirect.'">');
