+++ query.php
@@ -2,11 +2,7 @@
 <?php
 require 'functions.php';
 #Check if logged in and use read-only account if not
-if (isset($_SESSION["loggedIn"])){
-	$DB = new mysqli("localhost","momentu2_" . $_SESSION["user"],$_SESSION["pass"],"momentu2_frcteam4999");
-} else {
-	$DB = new mysqli("localhost","momentu2_ro","aRza#p=XckDC","momentu2_frcteam4999");
-}
+$DB = createDBObject();
 #Check if accessed by post and apply the filters if so
 if ($_SERVER["REQUEST_METHOD"] == "POST") {
 	$query = 'SELECT Team FROM '.getCurrentDB().' WHERE ';
