+++ specificvars.php.example
@@ -4,4 +4,14 @@ $DBUser = "";
 $DBPass = "";
 $LoginDBName = "";
 
+$Database = "";
+
+$LoginTableName = "";
+
+/*
+The user's username and password are checked against the database to see if they match. If they match, then the user has the LoggedIn flag set to true on their session.
+This causes createDBObject to check the user's username and password in the $LoginTableName, and if it is correct, the user is allowed to use the $DBUser and $DBPass to view and change the scouting information.
+*/
+
+
  ?>
